Add autoscale
Display is now able to autoscale. By default there are a few values
defined (0.2, 0.5, 1, 1.5 and 2 A) but you can adjust these by just
changing the array.

The drawing routine is now less efficient time-wise, since I now use
the previous values buffer with the read reading, and scale it to
24-0 at draw time. The reason for this, is that when scaling down,
this preserves the resolution of previous readings.

Please test!